Factor out the adapter interface
